Move scalar_* functions to separate header file
The scalar_* functions used for compression and decompression
are currently embedded in the poly.h file.
This commits moves them to a separate header file called scalar.h
and adjusts the CBMC proofs accordingly.

The motivation for this change is that we want to add native
implementations for the functions in verify.c.
However, including our native backend (native_arith.h) from verify.h
creates a circular dependency verify -> native_arith -> poly -> verify.
The only reason for including verify.h in poly.h is the need for
cmov_int16 in the scalar compression.
Moving it to a separate file resolves this dependency.

This cleaner as these functions are not present in the original
poly.h from the reference implementation.
