+++ cbmc/proofs/scalar_compress_d1/scalar_compress_d1_harness.c
@@ -16,7 +16,7 @@
  *   - include the declaration of the function
  *   - include the types needed to declare function arguments
  */
-#include <poly.h>
+#include <scalar.h>
 
 /**
  * @brief Starting point for formal analysis
