+++ cbmc/proofs/scalar_decompress_d5/scalar_decompress_d5_harness.c
@@ -16,7 +16,7 @@
  *   - include the declaration of the function
  *   - include the types needed to declare function arguments
  */
-#include <poly.h>
+#include <scalar.h>
 
 /**
  * @brief Starting point for formal analysis
