+++ cbmc/proofs/scalar_signed_to_unsigned_q/scalar_signed_to_unsigned_q_harness.c
@@ -17,7 +17,7 @@
  *   - include the declaration of the function
  *   - include the types needed to declare function arguments
  */
-#include <poly.h>
+#include <scalar.h>
 
 /**
  * @brief Starting point for formal analysis
